Deprecate '-api-tries' flag for new '-api-retries'
This choice of naming proved to be confusing for command line users and
also contributors. Instead of receiving number of attempts to be made,
switch to receiving number of retries, which is more customary.

If '-api-tries' flag is passed, its value - 1 gets used as the number of
retries and a deprecation warning is displayed.

This flag will be removed in v1.0.0 release.