Test request timeouts
TestPostRetries was not testing client request timeouts.
Co-opt 0 in the test tables as an indicator to go to sleep for twice the
length of client timeout, so actually test retry on timeout behavior.

To prevent this test becoming unreasonably long running due to default
1 second linear backoff step, add Backoff to APIClient struct to allow
configuration of this step duration.  Now the case can use 5ms as
backoff step, completing under 100ms.