Add the ability to truncate ping body
Adds `-ping-body-limit` flag to specify the tail size in bytes.
It defaults to 10KB (10,000 bytes), the default value of healthchecks.io
instance. A truncation notice will be appended to ping body, while
staying exactly at the limit. Passing 0 as the value disables
truncation.

For this feature we add and internal a ring buffer implementation to
truncate ping body to last N bytes without paying the memory cost of
retaining large amounts of stdout & stderr output in memory just to be
discarded.

Re: Feature Idea #9