Initial support for check slugs
Healthchecks is adding URL slugs support along with a per project ping
key as an alternative handle for checks.
https://twitter.com/healthchecks_io/status/1437679906633158665

Work in progress.

Cutting a beta for bug hunting and stabilizing the terminology with user
feedback.

Still need to write a unit test for business logic part of check handle
parameter precedence, and update the README.