Add support for log only pings
Healthchecks added a "log" ping type to cater to use cases where a
periodic command is likely to exit with failure status but expected to
eventually succeed in future instantiations.
https://github.com/healthchecks/healthchecks/issues/525#issuecomment-1139375529

Add `-log-only-non-zero` to send a log ping instead of a failure when the
command exits with a non-zero status. If runitor fails to execute the
command (e.g. not found, no permission, system ran out of pids, ...) a
failure ping is still sent.

Addresses #65