Add OS and arch information to user-agent
This updates the `User-Agent` HTTP header to include the operating
system and machine architecture of the runtime environment runitor is
running on.

Example:

Currently runitor sends:
  runitor/v1.2.0 (+https://bdd.fi/x/runitor)

After this change it will send an equivalent of:
  runitor/v1.2.0 (linux-amd64 +https://bdd.fi/x/runitor)

Who gets to see this header value?

If you are using healthchecks.io's hosted service, along with everything
sent in the ping, this header value is stored in their database and
possibly in the HTTP serving path logs.

If you are self hosting and the connection to your instance is over
HTTPS, then only you get to see it.