OCI Image: Allow reusing tag ver w/ newer builds
This change is for the maintainers only.

We want to update OCI images weekly to pick up underlying image updates.
This should not change the version. But when we need to pick up updates
in Go, we need to rebuild and publish with a new version (usually
vX.Y.Z-build.N), without a release tag.

- Add RELBUILD env var support to mkoci in addition to RELEASE to allow
  publishing RELEASE image using RELBUILD binaries from GitHub releases.

- `mkoci push` can now take a list of manifest names to push instead of
  building its own list from RELEASE and RUNTIMES.

- Start shipping major version images along with RELEASE and runtime
  image ones. e.g. `v1`, `v1-alpine`, ...