Containerfile: Fix scoping blues
A FROM directive nukes the ARGs defined by the previous FROM.  So the
ones that need to be global across stages need to be defined before the
first FROM.

TARGET... build args are already global but needs to be redefined empty
in a local scope. When using them in the FROM line, an attempt to
redefine them only makes is local to the prior scope. Hence the removal
of ARG TARGETPLATFORM before runtime image build.

While here, use docker.io/library/... instead of the legacy names.